Controle Ganhos

+++ main/filtro.cpp
@@ -1,10 +1,10 @@
 #include <stdio.h>
 #include <math.h>
 
-float filtro(float tam, float fc, float Ak, float Afko, float *Afk){
-    float pi = 3.1415926;
-    float T = tam/1000; //tempo de amostragem em segundos
-    float a1 = 1/(1+T*2*pi*fc), a2 = T*2*pi*fc/(1+T*2*pi*fc);
+int filtro(float tam, float fc, double Ak, double Afko, double *Afk){
+    double pi = 3.1415926;
+    double T = tam/1000; //tempo de amostragem em segundos
+    double a1 = 1/(1+T*2*pi*fc), a2 = T*2*pi*fc/(1+T*2*pi*fc);
     
     if ( fpclassify(Afko) == FP_NAN){
         Afko = 0;
