+++ main/filtro.h
@@ -1,5 +1,5 @@
 #ifndef FILTRO_H
 #define FILTRO_H
 
-int filtro(float tam, float fc, float Ak, float Afko, float *Afk);
+int filtro(float tam, float fc, double Ak, double Afko, double *Afk);
 #endif
